+++ frontend/components/location/searchPlaces.js
